create a new file related to cycle enumeration
